move command enrichments to parse.go
