+++ temporalcli/internal/cmd/gen-docs/main.go
@@ -32,11 +32,6 @@ func run() error {
 		return fmt.Errorf("failed parsing markdown: %w", err)
 	}
 
-	cmds, err = commandsgen.EnrichCommands(cmds)
-	if err != nil {
-		return fmt.Errorf("failed enriching commands: %w", err)
-	}
-
 	// Generate docs
 	b, err := commandsgen.GenerateDocsFiles(cmds)
 	if err != nil {
